redoing it... Still does not work...

+++ include/RandomContractionAlgorithm.h
@@ -6,7 +6,7 @@ class Graph;
 class RandomContractionAlgorithm
 {
 public:
-    static int run(Graph graph);
+    static size_t run(Graph graph);
 };
 
 #endif // RANDOMCONTRACTIONALGORITHM_H
